Revamp since the SAMLInitializer was moved to saml common repo
