+++ pom.xml
@@ -176,6 +176,13 @@
                 <version>${carbon.identity.framework.version}</version>
             </dependency>
 
+            <!--SAML Common Util dependency-->
+            <dependency>
+                <groupId>org.wso2.carbon.identity.saml.common</groupId>
+                <artifactId>org.wso2.carbon.identity.saml.common.util</artifactId>
+                <version>${saml.common.util.version}</version>
+            </dependency>
+
             <!-- Commons dependencies -->
             <dependency>
                 <groupId>commons-collections.wso2</groupId>
@@ -355,6 +362,10 @@
         <!--Carbon Identity Framework Version-->
         <carbon.identity.framework.version>5.14.120-SNAPSHOT</carbon.identity.framework.version>
         <carbon.identity.framework.imp.pkg.version.range>[5.14.67, 6.0.0)</carbon.identity.framework.imp.pkg.version.range>
+
+        <!--SAML Common Util Version-->
+        <saml.common.util.version>1.0.0-SNAPSHOT</saml.common.util.version>
+        <saml.common.util.version.range>[1.0.0,2.0.0)</saml.common.util.version.range>
 
         <!-- Carbon Commons -->
         <commons-collections.version>3.2.0.wso2v1</commons-collections.version>
